whiteboard-system/requirements.txt
openai>=1.0.0
pillow>=9.0.0
moviepy>=1.0.3
numpy>=1.21.0
python-dotenv>=0.19.0
